+++ src/main/java/racingcar/domian/Cars.java
@@ -7,31 +7,20 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 public class Cars {
-    private static final int CAR_LIMIT = 5;
-    private static final String DELIMITER = ",";
-    private static final String EMPTY_STRING = "";
 
+    private static final String EMPTY_STRING = "";
+    private static final CarFactory carFactory = new CarFactory();
     private final List<Car> cars;
 
     public Cars(String inputNames) {
-        String[] names = inputNames.split(DELIMITER);
-        validateCarCount(names.length);
-        List<Car> players = new ArrayList<>();
-        for (String name : names) {
-            players.add(new Car(name));
-        }
-        this.cars = players;
+        cars = carFactory.createCars(inputNames);
     }
 
     public List<Car> getCars() {
         return cars;
     }
 
-    private static void validateCarCount(int carCount) {
-        if (carCount > CAR_LIMIT) {
-            throw new IllegalArgumentException("자동차 5대 초과입니다!");
-        }
-    }
+
 
     private static boolean isNotEmpty(String maybeWinner) {
         return !EMPTY_STRING.equals(maybeWinner);
